IGAPP-336: move progressbar to common components
